Fixed some errors and warnings on teensy

+++ Atm_comparator.cpp
@@ -68,10 +68,10 @@ Atm_comparator &  Atm_comparator::average( uint16_t * v, uint16_t size )
    return *this;
 }
 
-Atm_comparator & Atm_comparator::bitmap( int v ) 
+Atm_comparator & Atm_comparator::bitmap( uint16_t v ) 
 {
    bitmap_sample = 0;
-   for ( int i = 0; i < p_threshold_size; i++ ) {
+   for ( uint16_t i = 0; i < p_threshold_size; i++ ) {
      if ( v >= p_threshold[i] ) bitmap_sample |= ( 1 << i );
    }
    bitmap_diff = bitmap_sample ^ bitmap_previous;
@@ -103,13 +103,13 @@ void Atm_comparator::action( int id )
   	  return;
   	case ACT_SEND :
       if ( v_sample >= v_previous ) {
-        for ( int i = 0; i < p_threshold_size; i++ ) {
+        for ( uint16_t i = 0; i < p_threshold_size; i++ ) {
           if ( (bitmap_diff >> i ) & 1 ) {
             (*callback)( v_sample, 1, i, p_threshold[i] );      
           }
         }        
       } else {
-        for ( int i = p_threshold_size; i >= 0; i-- ) {
+        for ( uint16_t i = p_threshold_size; i >= 0; i-- ) {
           if ( (bitmap_diff >> i ) & 1 ) {
             (*callback)( v_sample, 0, i, p_threshold[i] );      
           }
@@ -195,10 +195,10 @@ Att_comparator &  Att_comparator::average( uint16_t * v, uint16_t size )
    return *this;
 }
 
-Att_comparator & Att_comparator::bitmap( int v ) 
+Att_comparator & Att_comparator::bitmap( uint16_t v ) 
 {
    bitmap_sample = 0;
-   for ( int i = 0; i < p_threshold_size; i++ ) {
+   for ( uint16_t i = 0; i < p_threshold_size; i++ ) {
      if ( v >= p_threshold[i] ) bitmap_sample |= ( 1 << i );
    }
    bitmap_diff = bitmap_sample ^ bitmap_previous;
@@ -230,13 +230,13 @@ void Att_comparator::action( int id )
   	  return;
   	case ACT_SEND :
       if ( v_sample >= v_previous ) {
-        for ( int i = 0; i < p_threshold_size; i++ ) {
+        for ( uint16_t i = 0; i < p_threshold_size; i++ ) {
           if ( (bitmap_diff >> i ) & 1 ) {
             (*callback)( v_sample, 1, i, p_threshold[i] );      
           }
         }        
       } else {
-        for ( int i = p_threshold_size; i >= 0; i-- ) {
+        for ( uint16_t i = p_threshold_size; i >= 0; i-- ) {
           if ( (bitmap_diff >> i ) & 1 ) {
             (*callback)( v_sample, 0, i, p_threshold[i] );      
           }
